% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AscensorAndar.R
\name{AscensorAndar}
\alias{AscensorAndar}
\title{Regra do ascensor entre os andares}
\usage{
AscensorAndar(custo, nand)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}
}
\value{
Os distintos andares e a cantidade que lle corresponde pagar a cada un deles
}
\description{
Reparte o custo do ascensor entre os andares en base á regra do ascensor
}
\examples{
AscensorAndar(140,4)
AscensorAndar(120,5)
}
