% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AscensorApt.R
\name{AscensorApt}
\alias{AscensorApt}
\title{Reparto entre os apartamentos en base á regra do ascensor}
\usage{
AscensorApt(andar, custo, nand, napt)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}

\item{napt}{O número de apartamentos que ten o andar no que se atopa}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do ascensor. Emprega como unidade de reparto os apartamentos e como unións a priori os andares.
}
\examples{
AscensorApt(3,140,4,3)
AscensorApt(1,140,4,2)
}
