% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ascensorm2.R
\name{Ascensorm2}
\alias{Ascensorm2}
\title{Reparto entre os metros cadrados en base á regra do ascensor}
\usage{
Ascensorm2(andar, custo, nand, m2and, m2apt)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}

\item{m2and}{O número de metros cadrados que ten o andar no que se atopa}

\item{m2apt}{O número de metros cadrados que ten apartamento}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do ascensor. Emprega como unidade de reparto os metros cadrados e como unións a priori os andares.
}
\examples{
Ascensorm2(3,140,4,140,40)
Ascensorm2(1,140,4,150,60)
}
