% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REandar.R
\name{REandar}
\alias{REandar}
\title{Reparto equitativo entre os andares}
\usage{
REandar(custo, nand)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}
}
\value{
Os distintos andares e a cantidade que lle corresponde pagar a cada un deles
}
\description{
Reparte o custo do ascensor entre os andares en base á regra do Reparto Equitativo
}
\examples{
REandar(140,4)
REandar(120,5)
}
