% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REapt.R
\name{REapt}
\alias{REapt}
\title{Reparto equitativo entre os apartamentos}
\usage{
REapt(custo, napt)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{napt}{O número de apartamentos que ten o edificio}
}
\value{
A cantidade que lle corresponde pagar a cada apartamento
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do Reparto Equitativo. Emprega como unidade de reparto os apartamentos.
}
\examples{
REapt(140,7)
REapt(120,4)
}
