% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REaptU.R
\name{REaptU}
\alias{REaptU}
\title{Reparto equitativo entre os apartamentos con unións a priori}
\usage{
REaptU(custo, nand, aptand)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}

\item{aptand}{O número de apartamentos que ten o andar no que se atopa}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do Reparto Equitativo. Emprega como unidade de reparto os apartamentos e como unión a priori os andares.
}
\examples{
REaptU(140,4,2)
REaptU(140,4,3)
}
