% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REB1m2.R
\name{REB1m2}
\alias{REB1m2}
\title{Reparto entre os metros cadrados en base á regra REB1}
\usage{
REB1m2(andar, cbaixo, cand, nand, m2and, m2apt)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{cbaixo}{O custo correspondente aos traballos feitos no baixo}

\item{cand}{O custo correspondente aos traballos de cada andar adicional}

\item{nand}{O número de andares que ten o edificio}

\item{m2and}{O número de metros cadrados que ten o andar no que se atopa}

\item{m2apt}{O número de metros cadrados que ten apartamento}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra REB1. Emprega como unidade de reparto os metros cadrados e como unións a priori os andares.
}
\examples{
REB1m2(3,60,20,4,140,40)
REB1m2(1,60,20,4,150,60)
}
