% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REB3apt.R
\name{REB3apt}
\alias{REB3apt}
\title{Reparto entre os apartamentos en base á regra REB3}
\usage{
REB3apt(
  andar,
  aptand,
  cbaixo,
  cand,
  nand,
  napt1 = 0,
  napt2 = 0,
  napt3 = 0,
  napt4 = 0,
  napt5 = 0,
  napt6 = 0,
  napt7 = 0,
  napt8 = 0,
  napt9 = 0
)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{aptand}{O número de apartamentos do andar no que se atopa}

\item{cbaixo}{O custo correspondente aos traballos feitos no baixo}

\item{cand}{O custo correspondente aos traballos de cada andar adicional}

\item{nand}{O número de andares que ten o edificio}

\item{napt1}{O número de apartamentos que ten o andar 1}

\item{napt2}{O número de apartamentos que ten o andar 2}

\item{napt3}{O número de apartamentos que ten o andar 3}

\item{napt4}{O número de apartamentos que ten o andar 4}

\item{napt5}{O número de apartamentos que ten o andar 5}

\item{napt6}{O número de apartamentos que ten o andar 6}

\item{napt7}{O número de apartamentos que ten o andar 7}

\item{napt8}{O número de apartamentos que ten o andar 8}

\item{napt9}{O número de apartamentos que ten o andar 9}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra REB3. Emprega como unidade de reparto os apartamentos e como unións a priori os andares.
}
\examples{
REB3apt(3,3,60,20,4,2,1,3,1)
REB3apt(1,2,60,20,4,2,1,3,1)
}
