% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REB3m2.R
\name{REB3m2}
\alias{REB3m2}
\title{Reparto entre os metros cadrados en base á regra REB3}
\usage{
REB3m2(
  andar,
  m2and,
  cbaixo,
  cand,
  nand,
  m2apt,
  nm21 = 0,
  nm22 = 0,
  nm23 = 0,
  nm24 = 0,
  nm25 = 0,
  nm26 = 0,
  nm27 = 0,
  nm28 = 0,
  nm29 = 0
)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{m2and}{O número de metros cadrados do andar no que se atopa}

\item{cbaixo}{O custo correspondente aos traballos feitos no baixo}

\item{cand}{O custo correspondente aos traballos de cada andar adicional}

\item{nand}{O número de andares que ten o edificio}

\item{m2apt}{O número de metros cadrados que ten o apartamento}

\item{nm21}{O número de metros cadrados que ten o andar 1}

\item{nm22}{O número de metros cadrados que ten o andar 2}

\item{nm23}{O número de metros cadrados que ten o andar 3}

\item{nm24}{O número de metros cadrados que ten o andar 4}

\item{nm25}{O número de metros cadrados que ten o andar 5}

\item{nm26}{O número de metros cadrados que ten o andar 6}

\item{nm27}{O número de metros cadrados que ten o andar 7}

\item{nm28}{O número de metros cadrados que ten o andar 8}

\item{nm29}{O número de metros cadrados que ten o andar 9}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra REB3. Emprega como unidade de reparto os metros cadrados e como unións a priori os andares.
}
\examples{
REB3m2(3,140,60,20,4,40,150,150,140,150)
REB3m2(1,150,60,20,4,60,150,150,140,150)
}
