% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBm2.R
\name{REBm2}
\alias{REBm2}
\title{Reparto equitativo do beneficio entre os metros cadrados}
\usage{
REBm2(
  andar,
  cbaixo,
  cand,
  nand,
  m2apt,
  nm21 = 0,
  nm22 = 0,
  nm23 = 0,
  nm24 = 0,
  nm25 = 0,
  nm26 = 0,
  nm27 = 0,
  nm28 = 0,
  nm29 = 0
)
}
\arguments{
\item{andar}{O andar no que se atopa o apartamento}

\item{cbaixo}{O custo correspondente aos traballos feitos no baixo}

\item{cand}{O custo correspondente aos traballos de cada andar adicional}

\item{nand}{O número de andares que ten o edificio}

\item{m2apt}{O número de metros cadrados que ten o apartamento}

\item{nm21}{O número de metros cadrados que ten o andar 1}

\item{nm22}{O número de metros cadrados que ten o andar 2}

\item{nm23}{O número de metros cadrados que ten o andar 3}

\item{nm24}{O número de metros cadrados que ten o andar 4}

\item{nm25}{O número de metros cadrados que ten o andar 5}

\item{nm26}{O número de metros cadrados que ten o andar 6}

\item{nm27}{O número de metros cadrados que ten o andar 7}

\item{nm28}{O número de metros cadrados que ten o andar 8}

\item{nm29}{O número de metros cadrados que ten o andar 9}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do Reparto Equitativo do Beneficio. Emprega como unidade de reparto os metros cadrados.
}
\examples{
REBm2(3,60,20,4,40,150,150,140,150)
REBm2(1,60,20,4,60,150,150,140,150)
}
