% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REm2.R
\name{REm2}
\alias{REm2}
\title{Reparto equitativo entre os metros cadrados}
\usage{
REm2(custo, metros, m2apt)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{metros}{O número de metros cadrados que ten o edificio}

\item{m2apt}{O número de metros cadrados que ten o apartamento}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do Reparto Equitativo. Emprega como unidade de reparto os metros cadrados.
}
\examples{
REm2(140,590,40)
REm2(140,590,60)
}
