% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REm2U.R
\name{REm2U}
\alias{REm2U}
\title{Reparto equitativo entre os metros cadrados con unións a priori}
\usage{
REm2U(custo, nand, m2and, m2apt)
}
\arguments{
\item{custo}{O custo total da instalación do ascensor}

\item{nand}{O número de andares que ten o edificio}

\item{m2and}{O número de metros cadrados que ten o andar no que se atopa}

\item{m2apt}{O número de metros cadrados que ten apartamento}
}
\value{
A cantidade que lle corresponde pagar ao apartamento en cuestión
}
\description{
Calcula a cantidade a pagar do custo do ascensor por un apartamento en base á regra do Reparto Equitativo. Emprega como unidade de reparto os metros cadrados e como unión a priori os andares.
}
\examples{
REm2U(140,4,140,40)
REm2U(140,4,150,60)
}
